requirements.txt
langchain==0.3.25
langchain-openai==0.3.14

datasets==3.6.0
torch==2.7.0
torchaudio==2.7.0
transformers==4.52.1

TTS==0.22.0

sounddevice==0.5.1
soundfile==0.13.1
PyAudio==0.2.14
webrtcvad==2.0.10

diffusers==0.33.1
moviepy==2.1.2
opencv-python==4.11.0.86
pillow==10.4.0

requests==2.32.3
selenium==4.31.0
beautifulsoup4
yt-dlp

numpy==1.22.0
plyer==2.1.0
pooch==1.8.2
preshed==3.0.9
pygame==2.6.1
pyttsx3==2.98
dateparser==1.1.8